Update tasks.py
In this commit, the scheduling interval for the function bestchange_parse() has been modified. Previously, the function was set to execute every 5 minutes as per the BackgroundScheduler setup in the start_scheduler() function. However, this commit changes the interval from 5 minutes to 30 minutes, thereby reducing the frequency of the function's execution. This alteration might be in response to requirements for less frequent data parsing or to conserve system resources.