Create Telegram Page Component
In this commit, a new component named "TelegramPage" is created in the file telegram.tsx. The created component is a functional component using React Hooks. Currently, it only returns a div element containing a header with text "Telegram Page". This signifies the creation of a new page or feature related to Telegram in the project. The component is exported as the default export of the module.